Amp-Thread-ID: https://ampcode.com/threads/T-019fb458-7e27-73e2-90e8-6b15436827d2

+++ sdk/search/Azure.Search.Documents/tests/Samples/Sample07_VectorSearch_SemanticHybridSearch.cs
@@ -97,7 +97,10 @@ namespace Azure.Search.Documents.Tests.Samples.VectorSearch
             }
             finally
             {
-                await indexClient.DeleteIndexAsync(indexName, cancellationToken: CancellationToken.None);
+                if (indexClient != null)
+                {
+                    await indexClient.DeleteIndexAsync(indexName, cancellationToken: CancellationToken.None);
+                }
             }
         }
